refactor: rename file side panel layout
