+++ src/components/common/side-panel-layout.vue
