+++ src/components/layer-panel/layer-panel.vue
@@ -5,7 +5,7 @@ import { useTranslation } from 'i18next-vue'
 
 import CatalogTab from '@/components/catalog/catalog-tab.vue'
 import LayerManager from '@/components/layer-manager/layer-manager.vue'
-import SidePanelLayout from '@/components/common/panel-layout.vue'
+import SidePanelLayout from '@/components/common/side-panel-layout.vue'
 import { useAppStore } from '@/stores/app.store'
 import { useMapStore } from '@/stores/map.store'
 import useOffline from '@/composables/offline/offline.composable'
